+++ tiledb/sm/cpp_api/config.h
@@ -277,6 +277,11 @@ class Config {
    *    If `true`, an error will be thrown if there are cells with coordinates
    *    falling outside the array domain during sparse fragment writes. <br>
    *    **Default**: true
+   *    `sm.read_range_oob` <br>
+   *    If `error`, this will check ranges for read with out-of-bounds on the
+   *    dimension domain's and error. If `warn`, the ranges will be capped at
+   * the dimension's domain and a warning logged. <br>
+   *    **Default**: true
    * - `sm.check_global_order` <br>
    *    Checks if the coordinates obey the global array order. Applicable only
    *    to sparse writes in global order.